Refactor culling and visibility logic into VisibilitySystem
