+++ viewer3d/viewer3dcontroller.cpp
@@ -970,317 +970,10 @@ void Viewer3DController::SetSelectedUuids(
 }
 
 
-bool Viewer3DController::EnsureBoundsComputedImpl(
+bool Viewer3DController::EnsureBoundsComputed(
     const std::string &uuid, ItemType type,
     const std::unordered_set<std::string> &hiddenLayers) {
-  auto transformBounds = [](const Viewer3DController::BoundingBox &local,
-                            const Matrix &m) {
-    Viewer3DController::BoundingBox world;
-    world.min = {FLT_MAX, FLT_MAX, FLT_MAX};
-    world.max = {-FLT_MAX, -FLT_MAX, -FLT_MAX};
-    const auto &mn = local.min;
-    const auto &mx = local.max;
-    const std::array<std::array<float, 3>, 8> corners = {
-        std::array<float, 3>{mn[0], mn[1], mn[2]},
-        {mx[0], mn[1], mn[2]},
-        {mn[0], mx[1], mn[2]},
-        {mx[0], mx[1], mn[2]},
-        {mn[0], mn[1], mx[2]},
-        {mx[0], mn[1], mx[2]},
-        {mn[0], mx[1], mx[2]},
-        {mx[0], mx[1], mx[2]}};
-    for (const auto &c : corners) {
-      auto p = TransformPoint(m, c);
-      world.min[0] = std::min(world.min[0], p[0]);
-      world.min[1] = std::min(world.min[1], p[1]);
-      world.min[2] = std::min(world.min[2], p[2]);
-      world.max[0] = std::max(world.max[0], p[0]);
-      world.max[1] = std::max(world.max[1], p[1]);
-      world.max[2] = std::max(world.max[2], p[2]);
-    }
-    return world;
-  };
-
-  const auto &fixtures = SceneDataManager::Instance().GetFixtures();
-  const auto &trusses = SceneDataManager::Instance().GetTrusses();
-  const auto &objects = SceneDataManager::Instance().GetSceneObjects();
-
-  if (type == ItemType::Fixture) {
-    if (m_fixtureBounds.find(uuid) != m_fixtureBounds.end())
-      return true;
-    auto fit = fixtures.find(uuid);
-    if (fit == fixtures.end() ||
-        !IsLayerVisibleCached(hiddenLayers, fit->second.layer))
-      return false;
-
-    Viewer3DController::BoundingBox bb;
-    Matrix fix = fit->second.transform;
-    fix.o[0] *= RENDER_SCALE;
-    fix.o[1] *= RENDER_SCALE;
-    fix.o[2] *= RENDER_SCALE;
-    bool found = false;
-    bb.min = {FLT_MAX, FLT_MAX, FLT_MAX};
-    bb.max = {-FLT_MAX, -FLT_MAX, -FLT_MAX};
-
-    std::string gdtfPath;
-    auto gdtfIt = m_resolvedGdtfSpecs.find(ResolveCacheKey(fit->second.gdtfSpec));
-    if (gdtfIt != m_resolvedGdtfSpecs.end() && gdtfIt->second.attempted)
-      gdtfPath = gdtfIt->second.resolvedPath;
-    auto itg = m_loadedGdtf.find(gdtfPath);
-    if (itg != m_loadedGdtf.end()) {
-      auto bit = m_modelBounds.find(gdtfPath);
-      if (bit == m_modelBounds.end()) {
-        BoundingBox local;
-        local.min = {FLT_MAX, FLT_MAX, FLT_MAX};
-        local.max = {-FLT_MAX, -FLT_MAX, -FLT_MAX};
-        bool localFound = false;
-        for (const auto &obj : itg->second) {
-          for (size_t vi = 0; vi + 2 < obj.mesh.vertices.size(); vi += 3) {
-            std::array<float, 3> p = {obj.mesh.vertices[vi] * RENDER_SCALE,
-                                      obj.mesh.vertices[vi + 1] * RENDER_SCALE,
-                                      obj.mesh.vertices[vi + 2] * RENDER_SCALE};
-            p = TransformPoint(obj.transform, p);
-            local.min[0] = std::min(local.min[0], p[0]);
-            local.min[1] = std::min(local.min[1], p[1]);
-            local.min[2] = std::min(local.min[2], p[2]);
-            local.max[0] = std::max(local.max[0], p[0]);
-            local.max[1] = std::max(local.max[1], p[1]);
-            local.max[2] = std::max(local.max[2], p[2]);
-            localFound = true;
-          }
-        }
-        if (localFound)
-          bit = m_modelBounds.emplace(gdtfPath, local).first;
-      }
-      if (bit != m_modelBounds.end()) {
-        bb = transformBounds(bit->second, fix);
-        found = true;
-      }
-    }
-
-    if (!found) {
-      float half = 0.1f;
-      std::array<std::array<float, 3>, 8> corners = {
-          std::array<float, 3>{-half, -half, -half},
-          {half, -half, -half},
-          {-half, half, -half},
-          {half, half, -half},
-          {-half, -half, half},
-          {half, -half, half},
-          {-half, half, half},
-          {half, half, half}};
-      for (const auto &c : corners) {
-        auto p = TransformPoint(fix, c);
-        bb.min[0] = std::min(bb.min[0], p[0]);
-        bb.min[1] = std::min(bb.min[1], p[1]);
-        bb.min[2] = std::min(bb.min[2], p[2]);
-        bb.max[0] = std::max(bb.max[0], p[0]);
-        bb.max[1] = std::max(bb.max[1], p[1]);
-        bb.max[2] = std::max(bb.max[2], p[2]);
-      }
-    }
-    m_fixtureBounds[uuid] = bb;
-    return true;
-  }
-
-  if (type == ItemType::Truss) {
-    if (m_trussBounds.find(uuid) != m_trussBounds.end())
-      return true;
-    auto tit = trusses.find(uuid);
-    if (tit == trusses.end() || !IsLayerVisibleCached(hiddenLayers, tit->second.layer))
-      return false;
-
-    BoundingBox bb;
-    Matrix tm = tit->second.transform;
-    tm.o[0] *= RENDER_SCALE;
-    tm.o[1] *= RENDER_SCALE;
-    tm.o[2] *= RENDER_SCALE;
-    bool found = false;
-    bb.min = {FLT_MAX, FLT_MAX, FLT_MAX};
-    bb.max = {-FLT_MAX, -FLT_MAX, -FLT_MAX};
-
-    if (!tit->second.symbolFile.empty()) {
-      std::string path;
-      auto modelIt = m_resolvedModelRefs.find(ResolveCacheKey(tit->second.symbolFile));
-      if (modelIt != m_resolvedModelRefs.end() && modelIt->second.attempted)
-        path = modelIt->second.resolvedPath;
-      auto bit = m_modelBounds.find(path);
-      if (bit == m_modelBounds.end()) {
-        auto it = m_loadedMeshes.find(path);
-        if (it != m_loadedMeshes.end()) {
-          BoundingBox local;
-          local.min = {FLT_MAX, FLT_MAX, FLT_MAX};
-          local.max = {-FLT_MAX, -FLT_MAX, -FLT_MAX};
-          bool localFound = false;
-          for (size_t vi = 0; vi + 2 < it->second.vertices.size(); vi += 3) {
-            std::array<float, 3> p = {it->second.vertices[vi] * RENDER_SCALE,
-                                      it->second.vertices[vi + 1] * RENDER_SCALE,
-                                      it->second.vertices[vi + 2] * RENDER_SCALE};
-            local.min[0] = std::min(local.min[0], p[0]);
-            local.min[1] = std::min(local.min[1], p[1]);
-            local.min[2] = std::min(local.min[2], p[2]);
-            local.max[0] = std::max(local.max[0], p[0]);
-            local.max[1] = std::max(local.max[1], p[1]);
-            local.max[2] = std::max(local.max[2], p[2]);
-            localFound = true;
-          }
-          if (localFound)
-            bit = m_modelBounds.emplace(path, local).first;
-        }
-      }
-      if (bit != m_modelBounds.end()) {
-        bb = transformBounds(bit->second, tm);
-        found = true;
-      }
-    }
-
-    if (!found) {
-      float len = (tit->second.lengthMm > 0 ? tit->second.lengthMm : 1000.0f) * RENDER_SCALE;
-      float wid = (tit->second.widthMm > 0 ? tit->second.widthMm : 200.0f) * RENDER_SCALE;
-      float hgt = (tit->second.heightMm > 0 ? tit->second.heightMm : 200.0f) * RENDER_SCALE;
-      std::array<std::array<float, 3>, 8> corners = {
-          std::array<float, 3>{0.0f, -wid * 0.5f, 0.0f},
-          {len, -wid * 0.5f, 0.0f},
-          {0.0f, wid * 0.5f, 0.0f},
-          {len, wid * 0.5f, 0.0f},
-          {0.0f, -wid * 0.5f, hgt},
-          {len, -wid * 0.5f, hgt},
-          {0.0f, wid * 0.5f, hgt},
-          {len, wid * 0.5f, hgt}};
-      for (const auto &c : corners) {
-        auto p = TransformPoint(tm, c);
-        bb.min[0] = std::min(bb.min[0], p[0]);
-        bb.min[1] = std::min(bb.min[1], p[1]);
-        bb.min[2] = std::min(bb.min[2], p[2]);
-        bb.max[0] = std::max(bb.max[0], p[0]);
-        bb.max[1] = std::max(bb.max[1], p[1]);
-        bb.max[2] = std::max(bb.max[2], p[2]);
-      }
-    }
-    m_trussBounds[uuid] = bb;
-    return true;
-  }
-
-  if (m_objectBounds.find(uuid) != m_objectBounds.end())
-    return true;
-  auto oit = objects.find(uuid);
-  if (oit == objects.end() || !IsLayerVisibleCached(hiddenLayers, oit->second.layer))
-    return false;
-
-  BoundingBox bb;
-  Matrix tm = oit->second.transform;
-  bb.min = {FLT_MAX, FLT_MAX, FLT_MAX};
-  bb.max = {-FLT_MAX, -FLT_MAX, -FLT_MAX};
-  bool found = false;
-
-  if (!oit->second.geometries.empty()) {
-    for (const auto &geo : oit->second.geometries) {
-      std::string path;
-      auto modelIt = m_resolvedModelRefs.find(ResolveCacheKey(geo.modelFile));
-      if (modelIt != m_resolvedModelRefs.end() && modelIt->second.attempted)
-        path = modelIt->second.resolvedPath;
-      auto bit = m_modelBounds.find(path);
-      if (bit == m_modelBounds.end()) {
-        auto it = m_loadedMeshes.find(path);
-        if (it != m_loadedMeshes.end()) {
-          BoundingBox local;
-          local.min = {FLT_MAX, FLT_MAX, FLT_MAX};
-          local.max = {-FLT_MAX, -FLT_MAX, -FLT_MAX};
-          bool localFound = false;
-          for (size_t vi = 0; vi + 2 < it->second.vertices.size(); vi += 3) {
-            std::array<float, 3> p = {it->second.vertices[vi] * RENDER_SCALE,
-                                      it->second.vertices[vi + 1] * RENDER_SCALE,
-                                      it->second.vertices[vi + 2] * RENDER_SCALE};
-            local.min[0] = std::min(local.min[0], p[0]);
-            local.min[1] = std::min(local.min[1], p[1]);
-            local.min[2] = std::min(local.min[2], p[2]);
-            local.max[0] = std::max(local.max[0], p[0]);
-            local.max[1] = std::max(local.max[1], p[1]);
-            local.max[2] = std::max(local.max[2], p[2]);
-            localFound = true;
-          }
-          if (localFound)
-            bit = m_modelBounds.emplace(path, local).first;
-        }
-      }
-      if (bit == m_modelBounds.end())
-        continue;
-      Matrix geoTm = MatrixUtils::Multiply(tm, geo.localTransform);
-      geoTm.o[0] *= RENDER_SCALE;
-      geoTm.o[1] *= RENDER_SCALE;
-      geoTm.o[2] *= RENDER_SCALE;
-      BoundingBox geoWorld = transformBounds(bit->second, geoTm);
-      bb.min[0] = std::min(bb.min[0], geoWorld.min[0]);
-      bb.min[1] = std::min(bb.min[1], geoWorld.min[1]);
-      bb.min[2] = std::min(bb.min[2], geoWorld.min[2]);
-      bb.max[0] = std::max(bb.max[0], geoWorld.max[0]);
-      bb.max[1] = std::max(bb.max[1], geoWorld.max[1]);
-      bb.max[2] = std::max(bb.max[2], geoWorld.max[2]);
-      found = true;
-    }
-  } else if (!oit->second.modelFile.empty()) {
-    std::string path;
-    auto modelIt = m_resolvedModelRefs.find(ResolveCacheKey(oit->second.modelFile));
-    if (modelIt != m_resolvedModelRefs.end() && modelIt->second.attempted)
-      path = modelIt->second.resolvedPath;
-    auto bit = m_modelBounds.find(path);
-    if (bit == m_modelBounds.end()) {
-      auto it = m_loadedMeshes.find(path);
-      if (it != m_loadedMeshes.end()) {
-        BoundingBox local;
-        local.min = {FLT_MAX, FLT_MAX, FLT_MAX};
-        local.max = {-FLT_MAX, -FLT_MAX, -FLT_MAX};
-        bool localFound = false;
-        for (size_t vi = 0; vi + 2 < it->second.vertices.size(); vi += 3) {
-          std::array<float, 3> p = {it->second.vertices[vi] * RENDER_SCALE,
-                                    it->second.vertices[vi + 1] * RENDER_SCALE,
-                                    it->second.vertices[vi + 2] * RENDER_SCALE};
-          local.min[0] = std::min(local.min[0], p[0]);
-          local.min[1] = std::min(local.min[1], p[1]);
-          local.min[2] = std::min(local.min[2], p[2]);
-          local.max[0] = std::max(local.max[0], p[0]);
-          local.max[1] = std::max(local.max[1], p[1]);
-          local.max[2] = std::max(local.max[2], p[2]);
-          localFound = true;
-        }
-        if (localFound)
-          bit = m_modelBounds.emplace(path, local).first;
-      }
-    }
-    if (bit != m_modelBounds.end()) {
-      Matrix scaledTm = tm;
-      scaledTm.o[0] *= RENDER_SCALE;
-      scaledTm.o[1] *= RENDER_SCALE;
-      scaledTm.o[2] *= RENDER_SCALE;
-      bb = transformBounds(bit->second, scaledTm);
-      found = true;
-    }
-  }
-
-  if (!found) {
-    float half = 0.15f;
-    std::array<std::array<float, 3>, 8> corners = {
-        std::array<float, 3>{-half, -half, -half},
-        {half, -half, -half},
-        {-half, half, -half},
-        {half, half, -half},
-        {-half, -half, half},
-        {half, -half, half},
-        {-half, half, half},
-        {half, half, half}};
-    for (const auto &c : corners) {
-      auto p = TransformPoint(tm, c);
-      bb.min[0] = std::min(bb.min[0], p[0]);
-      bb.min[1] = std::min(bb.min[1], p[1]);
-      bb.min[2] = std::min(bb.min[2], p[2]);
-      bb.max[0] = std::max(bb.max[0], p[0]);
-      bb.max[1] = std::max(bb.max[1], p[1]);
-      bb.max[2] = std::max(bb.max[2], p[2]);
-    }
-  }
-
-  m_objectBounds[uuid] = bb;
-  return true;
+  return m_visibilitySystem->EnsureBoundsComputed(uuid, type, hiddenLayers);
 }
 
 // Loads meshes or GDTF models referenced by scene objects. Called when the
@@ -1925,215 +1618,27 @@ void Viewer3DController::UpdateResourcesIfDirty() {
 bool Viewer3DController::TryBuildLayerVisibleCandidates(
     const std::unordered_set<std::string> &hiddenLayers,
     VisibleSet &out) const {
-  const auto &sceneObjects = SceneDataManager::Instance().GetSceneObjects();
-  const auto &trusses = SceneDataManager::Instance().GetTrusses();
-  const auto &fixtures = SceneDataManager::Instance().GetFixtures();
-
-  std::lock_guard<std::mutex> lock(m_sortedListsMutex);
-  out.objectUuids.clear();
-  out.trussUuids.clear();
-  out.fixtureUuids.clear();
-
-  out.objectUuids.reserve(m_sortedObjects.size());
-  for (const auto *entry : m_sortedObjects) {
-    if (!entry)
-      continue;
-    const auto &uuid = entry->first;
-    const auto &obj = entry->second;
-    if (!IsLayerVisibleCached(hiddenLayers, obj.layer))
-      continue;
-    if (sceneObjects.find(uuid) != sceneObjects.end())
-      out.objectUuids.push_back(uuid);
-  }
-
-  out.trussUuids.reserve(m_sortedTrusses.size());
-  for (const auto *entry : m_sortedTrusses) {
-    if (!entry)
-      continue;
-    const auto &uuid = entry->first;
-    const auto &truss = entry->second;
-    if (!IsLayerVisibleCached(hiddenLayers, truss.layer))
-      continue;
-    if (trusses.find(uuid) != trusses.end())
-      out.trussUuids.push_back(uuid);
-  }
-
-  out.fixtureUuids.reserve(m_sortedFixtures.size());
-  for (const auto *entry : m_sortedFixtures) {
-    if (!entry)
-      continue;
-    const auto &uuid = entry->first;
-    const auto &fixture = entry->second;
-    if (!IsLayerVisibleCached(hiddenLayers, fixture.layer))
-      continue;
-    if (fixtures.find(uuid) != fixtures.end())
-      out.fixtureUuids.push_back(uuid);
-  }
-
-  return true;
+  return m_visibilitySystem->TryBuildLayerVisibleCandidates(hiddenLayers, out);
 }
 
-bool Viewer3DController::TryBuildVisibleSetImpl(
-    const ViewFrustumSnapshot &frustum,
-    bool useFrustumCulling, float minPixels,
-    const VisibleSet &layerVisibleCandidates, VisibleSet &out) const {
-  out.objectUuids.clear();
-  out.trussUuids.clear();
-  out.fixtureUuids.clear();
-
-  out.objectUuids.reserve(layerVisibleCandidates.objectUuids.size());
-  for (const auto &uuid : layerVisibleCandidates.objectUuids) {
-    if (useFrustumCulling) {
-      auto bit = m_objectBounds.find(uuid);
-      if (bit == m_objectBounds.end())
-        continue;
-      ScreenRect rect;
-      bool anyDepthVisible = false;
-      if (!ProjectBoundingBoxToScreen(bit->second.min, bit->second.max,
-                                      frustum.viewport[3], frustum.model,
-                                      frustum.projection, frustum.viewport,
-                                      rect, anyDepthVisible) ||
-          !anyDepthVisible ||
-          ShouldCullByScreenRect(rect, frustum.viewport[2], frustum.viewport[3],
-                                 minPixels)) {
-        continue;
-      }
-    }
-    out.objectUuids.push_back(uuid);
-  }
-
-  out.trussUuids.reserve(layerVisibleCandidates.trussUuids.size());
-  for (const auto &uuid : layerVisibleCandidates.trussUuids) {
-    if (useFrustumCulling) {
-      auto bit = m_trussBounds.find(uuid);
-      if (bit == m_trussBounds.end())
-        continue;
-      ScreenRect rect;
-      bool anyDepthVisible = false;
-      if (!ProjectBoundingBoxToScreen(bit->second.min, bit->second.max,
-                                      frustum.viewport[3], frustum.model,
-                                      frustum.projection, frustum.viewport,
-                                      rect, anyDepthVisible) ||
-          !anyDepthVisible ||
-          ShouldCullByScreenRect(rect, frustum.viewport[2], frustum.viewport[3],
-                                 minPixels)) {
-        continue;
-      }
-    }
-    out.trussUuids.push_back(uuid);
-  }
-
-  out.fixtureUuids.reserve(layerVisibleCandidates.fixtureUuids.size());
-  for (const auto &uuid : layerVisibleCandidates.fixtureUuids) {
-    if (useFrustumCulling) {
-      auto bit = m_fixtureBounds.find(uuid);
-      if (bit == m_fixtureBounds.end())
-        continue;
-      ScreenRect rect;
-      bool anyDepthVisible = false;
-      if (!ProjectBoundingBoxToScreen(bit->second.min, bit->second.max,
-                                      frustum.viewport[3], frustum.model,
-                                      frustum.projection, frustum.viewport,
-                                      rect, anyDepthVisible) ||
-          !anyDepthVisible ||
-          ShouldCullByScreenRect(rect, frustum.viewport[2], frustum.viewport[3],
-                                 minPixels)) {
-        continue;
-      }
-    }
-    out.fixtureUuids.push_back(uuid);
-  }
-
-  return true;
+bool Viewer3DController::TryBuildVisibleSet(
+    const ViewFrustumSnapshot &frustum, bool useFrustumCulling,
+    float minPixels, const VisibleSet &layerVisibleCandidates,
+    VisibleSet &out) const {
+  return m_visibilitySystem->TryBuildVisibleSet(
+      frustum, useFrustumCulling, minPixels, layerVisibleCandidates, out);
 }
 
 const Viewer3DController::VisibleSet &Viewer3DController::GetVisibleSet(
     const ViewFrustumSnapshot &frustum,
     const std::unordered_set<std::string> &hiddenLayers,
     bool useFrustumCulling, float minPixels) const {
-  const bool layerCandidatesCacheValid =
-      (m_layerVisibleCandidatesSceneVersion == m_sceneVersion) &&
-      (m_layerVisibleCandidatesHiddenLayers == hiddenLayers);
-
-  if (!layerCandidatesCacheValid) {
-    VisibleSet builtCandidates;
-    if (TryBuildLayerVisibleCandidates(hiddenLayers, builtCandidates)) {
-      m_cachedLayerVisibleCandidates = std::move(builtCandidates);
-      m_layerVisibleCandidatesSceneVersion = m_sceneVersion;
-      m_layerVisibleCandidatesHiddenLayers = hiddenLayers;
-      ++m_layerVisibleCandidatesRevision;
-    }
-  }
-
-  const bool sameViewport = std::equal(
-      std::begin(frustum.viewport), std::end(frustum.viewport),
-      m_visibleSetViewport.begin());
-  const bool sameModel =
-      std::equal(std::begin(frustum.model), std::end(frustum.model),
-                 m_visibleSetModel.begin());
-  const bool sameProjection =
-      std::equal(std::begin(frustum.projection), std::end(frustum.projection),
-                 m_visibleSetProjection.begin());
-
-  const bool cacheValid =
-      (m_visibleSetLayerCandidatesRevision == m_layerVisibleCandidatesRevision) &&
-      (m_visibleSetFrustumCulling == useFrustumCulling) &&
-      (m_visibleSetMinPixels == minPixels) && sameViewport && sameModel &&
-      sameProjection;
-
-  if (!cacheValid) {
-    if (m_cachedLayerVisibleCandidates.Empty()) {
-      m_cachedVisibleSet = {};
-      m_visibleSetLayerCandidatesRevision = m_layerVisibleCandidatesRevision;
-      m_visibleSetFrustumCulling = useFrustumCulling;
-      m_visibleSetMinPixels = minPixels;
-      std::copy(std::begin(frustum.viewport), std::end(frustum.viewport),
-                m_visibleSetViewport.begin());
-      std::copy(std::begin(frustum.model), std::end(frustum.model),
-                m_visibleSetModel.begin());
-      std::copy(std::begin(frustum.projection), std::end(frustum.projection),
-                m_visibleSetProjection.begin());
-      return m_cachedVisibleSet;
-    }
-
-    VisibleSet built;
-    if (TryBuildVisibleSet(frustum, useFrustumCulling, minPixels,
-                           m_cachedLayerVisibleCandidates, built)) {
-      m_cachedVisibleSet = std::move(built);
-      m_visibleSetLayerCandidatesRevision = m_layerVisibleCandidatesRevision;
-      m_visibleSetFrustumCulling = useFrustumCulling;
-      m_visibleSetMinPixels = minPixels;
-      std::copy(std::begin(frustum.viewport), std::end(frustum.viewport),
-                m_visibleSetViewport.begin());
-      std::copy(std::begin(frustum.model), std::end(frustum.model),
-                m_visibleSetModel.begin());
-      std::copy(std::begin(frustum.projection), std::end(frustum.projection),
-                m_visibleSetProjection.begin());
-    }
-  }
-
-  return m_cachedVisibleSet;
+  return m_visibilitySystem->GetVisibleSet(frustum, hiddenLayers,
+                                           useFrustumCulling, minPixels);
 }
 
-void Viewer3DController::RebuildVisibleSetCacheImpl() {
-  ConfigManager &cfg = ConfigManager::Get();
-  const auto hiddenLayers = SnapshotHiddenLayers(cfg);
-  const CullingSettings culling = GetCullingSettings3D(cfg);
-  int viewport[4] = {0, 0, 0, 0};
-  double model[16] = {0.0};
-  double proj[16] = {0.0};
-  glGetIntegerv(GL_VIEWPORT, viewport);
-  glGetDoublev(GL_MODELVIEW_MATRIX, model);
-  glGetDoublev(GL_PROJECTION_MATRIX, proj);
-
-  ViewFrustumSnapshot frustum{};
-  std::copy(std::begin(viewport), std::end(viewport),
-            std::begin(frustum.viewport));
-  std::copy(std::begin(model), std::end(model), std::begin(frustum.model));
-  std::copy(std::begin(proj), std::end(proj), std::begin(frustum.projection));
-
-  const float minCullingPixels = culling.minPixels3D;
-  (void)GetVisibleSet(frustum, hiddenLayers, culling.enabled, minCullingPixels);
+void Viewer3DController::RebuildVisibleSetCache() {
+  m_visibilitySystem->RebuildVisibleSetCache();
 }
 
 // Renders all scene objects using their transformMatrix
@@ -4543,24 +4048,6 @@ void Viewer3DController::SetLayerColor(const std::string &layer,
 std::shared_ptr<const SymbolDefinitionSnapshot>
 Viewer3DController::GetBottomSymbolCacheSnapshot() const {
   return m_bottomSymbolCache.Snapshot();
-}
-
-bool Viewer3DController::EnsureBoundsComputed(
-    const std::string &uuid, ItemType type,
-    const std::unordered_set<std::string> &hiddenLayers) {
-  return m_visibilitySystem->EnsureBoundsComputed(uuid, type, hiddenLayers);
-}
-
-bool Viewer3DController::TryBuildVisibleSet(
-    const ViewFrustumSnapshot &frustum, bool useFrustumCulling,
-    float minPixels, const VisibleSet &layerVisibleCandidates,
-    VisibleSet &out) const {
-  return m_visibilitySystem->TryBuildVisibleSet(
-      frustum, useFrustumCulling, minPixels, layerVisibleCandidates, out);
-}
-
-void Viewer3DController::RebuildVisibleSetCache() {
-  m_visibilitySystem->RebuildVisibleSetCache();
 }
 
 void Viewer3DController::DrawMeshWithOutline(
